+++ R/cfrnow-package.R
@@ -6,6 +6,7 @@
 #' @importFrom epidist assert_epidist epidist_family_model epidist_formula_model
 #' @importFrom epidist epidist_model_prior epidist_stancode
 #' @importFrom epidist epidist_transform_data_model
+#' @importFrom rlang .data
 #' @importFrom stats plogis quantile rlnorm rnorm runif setNames
 ## usethis namespace: end
 NULL